refactor(ApiService): Create a new instance instead of a singleton
In long running processes, such as a queue worker, the JWT can expire.
The simplest fix right now is to create a new class every time we need a
new ApiService. If this turns into a memory issue, the fix would be to
create a new Guzzle Client for each ApiService and retry the request if
a request fails due to authentication.